+++ locale/localization.enUS.lua
@@ -12,6 +12,8 @@ local L = HRT.Localization
 L["options.general"] = "General Options"
 L["options.general.notification.name"] = "Chat Notification"
 L["options.general.notification.tooltip"] = "Activate or deactivate the notification in the chat after a combat."
+L["options.general.minimap-button.name"] = "Minimap Button"
+L["options.general.minimap-button.tooltip"] = "When this is enabled, the minimap button is displayed."
 
 L["options.combat-time-tracker"] = "Combat Time Tracker"
 L["options.combat-time-tracker.scale.name"] = "UI Scale"
@@ -28,24 +30,22 @@ L["options.about.game-version"] = "Game Version"
 L["options.about.addon-version"] = "Addon Version"
 L["options.about.author"] = "Author"
 
-L["options.about.button-delete-data.name"] = "Delete Combat Data"
-L["options.about.button-delete-data.tooltip"] = "Deletes all combat data that had been recorded and saved up to that point. This includes best times and how often a boss was defeated."
-L["options.about.button-delete-data.button"] = "Delete"
+L["options.about.button-delete-data.name"] = "???"
+L["options.about.button-delete-data.tooltip"] = "???"
+L["options.about.button-delete-data.button"] = "???"
 
 L["options.about.button-github.name"] = "Feedback & Help"
 L["options.about.button-github.tooltip"] = "Opens a popup window with a link to GitHub."
 L["options.about.button-github.button"] = "GitHub"
 
--- Combat Time Tracker
+-- General
 
-L["combat-time-tracker.unknown"] = "unknown"
-L["combat-time-tracker.button-reset"] = "Reset"
-L["combat-time-tracker.wait-combat"] = "Waiting for combat..."
+L["minimap-button.tooltip"] = "|cnLINK_FONT_COLOR:Left-click|r to show or hide the |cnGOLD_FONT_COLOR:Combat Time Tracker|r.\n|cnLINK_FONT_COLOR:Right-click|r to open the options."
 
 -- Dialog
 
 L["dialog.copy-address.text"] = "To copy the link press CTRL + C."
-L["dialog.delete-data.text"] = "Do you really want to delete all combat data?\n|cnNORMAL_FONT_COLOR:Warning:|r The game interface will be automatically reloaded!"
+L["dialog.delete-data.text"] = "Do you really want to delete all addon data?\n|cnNORMAL_FONT_COLOR:Warning:|r The game interface will be automatically reloaded!"
 
 -- Chat
 
@@ -56,3 +56,9 @@ L["chat.first-victory"] = "This was your first victory over |cnGOLD_FONT_COLOR:%
 L["chat.another-victory"] = "You have already defeated |cnGOLD_FONT_COLOR:%s|r (%s) once. (Victories: %s - Wipes: %s)"
 L["chat.first-wipe"] = "This was your first defeat against |cnGOLD_FONT_COLOR:%s|r (%s). (Victories: %s)"
 L["chat.another-wipe"] = "You have already lost to |cnGOLD_FONT_COLOR:%s|r (%s) once. (Victories: %s - Wipes: %s)"
+
+-- Combat Time Tracker
+
+L["combat-time-tracker.unknown"] = "unknown"
+L["combat-time-tracker.button-reset"] = "Reset"
+L["combat-time-tracker.wait-combat"] = "Waiting for combat..."