+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Horatum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Horatum/releases)
 
+**Important note: All settings have been reset to their default values!**
+
+- Added: Minimap Button
+- Changed: Consolidation of the options menu
+- Updated: deDE, enUS localization
 - Minor code adjustments